Adapt tiling mode depending on gfx driver.
So the Mesa gallium radeonsi driver for OpenGL on AMD
gpu's on Linux ignores the GL_TILING_MODE_EXT setting
for textures -- it always uses linear tiling. At least
as of Mesa 19.0.8. Which means on Linux + Mesa we must
use linear tiling on the Vulkan export side to make it
work.

Otoh. at least the NVidia binary blob on Linux, and
presumably on Windows, needs optimal tiling, or it
won't work at all, as render-to-texture on NVidia
apparently only works with tiled formats, not linear,
so we need optimal tiling for at least NVidia.

Assume for the moment that optimal tiling works
everywhere except on Mesa AMD, so select that
if not on Mesa, and linear on Mesa.